Minor Fixes #15 - Changed post http request to delete - Renamed files to userRoute.js and catalogRoute.js - Renamed functions in userRoute.js and catalogRoute.js
